im just commiting this
- So this works for googleTrans
- Im basically though saying
   - We only deal with wav data with no headers from synth_to_bytes
But this causes real issues down the line. Take elevelabs. We would be ditching the idea of getting mp3 data or other.

I might be ok with that - but I might not - i cant decide. Its a ton of work to strip out the format stuff from each engine though